Fix styling for comment box
